changes from review
