+++ app/components/PlayerManager.tsx
@@ -278,6 +278,10 @@ function PlayerManager({
 
   const value: PlayerSelection = {
     selectSource,
+    // Expose a simple way to load a demo bag for first launch onboarding.
+    // In the future we may want to replace this limited API with something more cohesive
+    // that exposes the different buildPlayerFromX methods above. At the same time,
+    // the prompt() responsibilities could be moved out of the PlayerManager.
     setPlayerFromDemoBag: async () =>
       buildPlayer(await buildPlayerFromBagURLs([DEMO_BAG_URL], buildPlayerOptions)),
     availableSources: playerSources,
